test(WinningNumberTest): separate tests for drawsystem
separate DrawMachineTest and WinningNumberTest
